requirements.txt
alembic==1.7.7
blinker==1.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.0
itsdangerous==2.0.0
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.1
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.32
stripe==3.2.0
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
wtf==0.1
WTForms==3.0.1
zipp==3.8.0
